Move LogCaptureAppender to common package in clients module
