+++ clients/src/test/java/org/apache/kafka/common/utils/LogCaptureAppender.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.streams.processor.internals.testutil;
+package org.apache.kafka.common.utils;
 
 import org.apache.log4j.AppenderSkeleton;
 import org.apache.log4j.Level;
